+++ src/main/java/demo/codeexample/project/domain/Project.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.time.LocalDate;
+import java.util.Set;
 
 @Getter
 @Setter
@@ -16,26 +17,25 @@ public class Project {
     private String title;
     private String description;
     private LocalDate releaseDate;
-    private final Long producerId;
+    private final Set<Long> employeesId;
     private final Category category;
     private Genre genre;
     private String imageURL;
 
-    public Project(Long id, String title, String description, LocalDate releaseDate, Long producerId,
+    public Project(Long id, String title, String description, LocalDate releaseDate, Set<Long> employeesId,
                    Category category, Genre genre, String imageURL) {
         this.id = id;
         this.title = title;
         this.description = description;
         this.releaseDate = releaseDate;
-        this.producerId = producerId;
+        this.employeesId = employeesId;
         this.category = category;
         this.genre = genre;
         this.imageURL = imageURL;
     }
 
-    public static Project createNew(String title, String description, LocalDate releaseDate, Long producerId,
+    public static Project createNew(String title, String description, LocalDate releaseDate, Set<Long> employeesId,
                                     Category category, Genre genre, String imageURL){
-        return new Project(null, title, description, releaseDate, producerId, category, genre, imageURL);
+        return new Project(null, title, description, releaseDate, employeesId, category, genre, imageURL);
     }
-
 }
