+++ src/main/java/demo/codeexample/user/CreateUserRequestDTO.java
@@ -1,5 +1,6 @@
 package demo.codeexample.user;
 
+import demo.codeexample.shared.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
