+++ src/main/java/demo/codeexample/user/domain/User.java
@@ -1,6 +1,6 @@
 package demo.codeexample.user.domain;
 
-import demo.codeexample.user.Role;
+import demo.codeexample.shared.Role;
 import jakarta.persistence.*;
 import lombok.*;
