+++ src/main/java/demo/codeexample/user/web/UserController.java
@@ -4,7 +4,7 @@ import demo.codeexample.exceptions.UserNotFoundException;
 import demo.codeexample.user.CreateUserRequestDTO;
 import demo.codeexample.user.UserDto;
 import demo.codeexample.user.application.UserService;
-import demo.codeexample.user.Role;
+import demo.codeexample.shared.Role;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
